Refactor Navigators. Add index file to export nested components
