+++ src/navigators/AppNavigator.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { DrawerNavigator } from 'react-navigation';
 
 import DashboardScreen from '../components/dashboard/DashboardScreen';
-import TournamentNavigator from '../navigators/TournamentNavigator';
+import { TournamentNavigator } from './Tournaments';
 import LeftMenu from '../components/LeftMenu';
 
 const routeConfiguration = {
@@ -15,4 +15,6 @@ const drawerConfiguration = {
   contentComponent: props => <LeftMenu {...props} />
 };
 
-export default AppNavigator = DrawerNavigator(routeConfiguration, drawerConfiguration);
+const AppNavigator = DrawerNavigator(routeConfiguration, drawerConfiguration);
+
+export default AppNavigator;
