+++ src/navigators/Tournaments/TournamentDetailNavigator.js
@@ -7,10 +7,9 @@ import {
   Text
 } from 'native-base';
 
-
-import FixtureScreen from '../components/tournaments/Fixture/FixtureScreen';
-import PositionsScreen from '../components/tournaments/Positions/PositionsScreen';
-import ScoresScreen from '../components/tournaments/ScoresScreen';
+import FixtureScreen from '../../components/tournaments/Fixture/FixtureScreen';
+import PositionsScreen from '../../components/tournaments/Positions/PositionsScreen';
+import ScoresScreen from '../../components/tournaments/ScoresScreen';
 
 const routeConfiguration = {
     Fixture: {
@@ -22,7 +21,7 @@ const routeConfiguration = {
     Scores: {
       screen: ScoresScreen
     }
-}
+};
 
 const tabBarConfiguration = {
   initialRouteName: 'Fixture',
@@ -38,7 +37,7 @@ const tabBarConfiguration = {
             vertical
             active={props.navigationState.index === 0}
             onPress={() => {
-              props.navigation.navigate('Fixture', { id, title })
+              props.navigation.navigate('Fixture', { id, title });
             }}
           >
             <Text>Fixture</Text>
@@ -47,7 +46,7 @@ const tabBarConfiguration = {
             vertical
             active={props.navigationState.index === 1}
             onPress={() => {
-              props.navigation.navigate('Positions', { id, title })
+              props.navigation.navigate('Positions', { id, title });
             }}
           >
             <Text>Positions</Text>
@@ -63,7 +62,8 @@ const tabBarConfiguration = {
       </Footer>
     );
   }
-}
+};
 
-export default TournamentDetailNavigator =
-  TabNavigator(routeConfiguration, tabBarConfiguration);
+const TournamentDetailNavigator = TabNavigator(routeConfiguration, tabBarConfiguration);
+  
+export default TournamentDetailNavigator;
