+++ src/navigators/Tournaments/TournamentNavigator.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { StackNavigator } from 'react-navigation';
 
-import TournamentListScreen from '../components/tournaments/TournamentListScreen';
+import TournamentListScreen from '../../components/tournaments/TournamentListScreen';
 import TournamentDetailNavigator from './TournamentDetailNavigator';
-import Toolbar from '../components/Toolbar';
+import Toolbar from '../../components/Toolbar';
 
 const routeConfiguration = {
   TournamentList: {
@@ -34,4 +34,9 @@ const routeConfiguration = {
   }
 };
 
-export default TournamentNavigator = StackNavigator(routeConfiguration, { initialRouteName: 'TournamentList' });
+const TournamentNavigator = StackNavigator(
+  routeConfiguration, 
+  { initialRouteName: 'TournamentList' }
+);
+
+export default TournamentNavigator;
