+++ src/navigators/Tournaments/index.js
@@ -1,0 +1,5 @@
+import TournamentNavigator from './TournamentNavigator';
+
+export {
+  TournamentNavigator,
+};
