+++ src/navigators/index.js
@@ -1,0 +1,7 @@
+import AppNavigator from './AppNavigator';
+import { TournamentNavigator } from './Tournaments';
+
+export {
+  AppNavigator,
+  TournamentNavigator,
+};
